argoslabs/data/excelread/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# for Excel
openpyxl

# for csv
xlwings==0.23.0




